Post Go-Live updates. Contains fixes for session management.
Outstanding issue: When the RSA Access manager 8 server times-out a session, the connector tries to establish a new connection. This first attempt fails and throws an RSA exception. It is not clear yet why RSA throws this exception, and RSA has been engaged to try to determine the cause of this exception.

In the meantime, the connector code has been modified to extablish a new session before issuing every command. Although this adds overhead and network traffic, it works everytime and doesn't seem to cause any noticeable latency. We will update the connector's session management once we hear back from RSA.

alex @ nulli . com